backend/requirements.txt
Django>=5.0
djangorestframework
django-cors-headers
psycopg2-binary
djangorestframework-simplejwt
python-dotenv
whitenoise
gunicorn
dj-database-url
